+++ Elements/test/SolidTests.cs
@@ -210,12 +210,12 @@ namespace Elements.Tests
             var F = new Vector3(1, 0, 1);
             var G = new Vector3(1, 1, 1);
             var H = new Vector3(0, 1, 1);
-            solid.AddFace(new Polygon(new[] { A, B, C, D }));
-            solid.AddFace(new Polygon(new[] { E, F, G, H }));
-            solid.AddFace(new Polygon(new[] { A, B, F, E }));
-            solid.AddFace(new Polygon(new[] { B, C, G, F }));
-            solid.AddFace(new Polygon(new[] { C, D, H, G }));
-            solid.AddFace(new Polygon(new[] { D, A, E, H }));
+            solid.AddFace(new Polygon(new[] { A, B, C, D }), mergeVerticesAndEdges: true);
+            solid.AddFace(new Polygon(new[] { E, F, G, H }), mergeVerticesAndEdges: true);
+            solid.AddFace(new Polygon(new[] { A, B, F, E }), mergeVerticesAndEdges: true);
+            solid.AddFace(new Polygon(new[] { B, C, G, F }), mergeVerticesAndEdges: true);
+            solid.AddFace(new Polygon(new[] { C, D, H, G }), mergeVerticesAndEdges: true);
+            solid.AddFace(new Polygon(new[] { D, A, E, H }), mergeVerticesAndEdges: true);
 
             var emptySolid = new ConstructedSolid(new Solid(), false);
             var import = new ConstructedSolid(solid, false);