+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `ConstructedSolid` serializes and deserializes correctly.
+- `Solid.AddFace(Polygon, Polygon[])` can take an optional third `mergeVerticesAndEdges` argument which will automatically reuse existing edges + vertices in the solid.
 
 ## 0.8.5